ip_tracking_project/requirements.txt
# Django and core dependencies
Django>=5.0,<6.0
django-ipware>=5.0.0
django-ratelimit>=4.1.0
djangorestframework>=3.14.0

# Production server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Task queue and caching
celery>=5.2.0
redis>=4.5.0

# IP Geolocation
geoip2>=4.7.0
maxminddb>=2.6.0
maxminddb-geolite2>=2018.703

# API Documentation
drf-yasg>=1.21.7
